requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
PySocks==1.7.1
python-datetime-tz==0.5.4
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
tornado==6.2
urllib3==1.26.12
